+++ src/components/Select/CHANGELOG.md
@@ -6,3 +6,5 @@
 - 2025-10-11: Support `id` prop for better form integration. Make it be able to
   uncontrolled. Support `defaultValue` prop. Code formatting improvements.
 - 2025-10-26: Update styles (padding) to make it look better.
+- 2025-10-27: Update styles when Select is opened or focused. Make its
+  `min-width` be `8rem` now (rather than `16rem`).